backend/requirements.txt
#REQUERIMENTOS DEL BACKEND
#pip install -r requirements.txt
aiodns==3.5.0
aiohttp==3.11.10
aiomultiprocess==0.9.1
aiosqlite==0.21.0
alembic==1.16.2
beautifulsoup4==4.13.4
censys==2.2.17
certifi==2024.8.30
Flask==3.1.1
flask_cors==6.0.1
flask_migrate==4.1.0
flask_sqlalchemy==3.1.1
netaddr==1.3.0
playwright==1.53.0
pydantic==2.11.7
pytest==8.4.1
python-dotenv==1.1.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.4
shodan==1.31.0
slowapi==0.1.9
SQLAlchemy==2.0.41
ujson==5.10.0
gunicorn==23.0.0
psycopg2-binary==2.9.10